bigquery/requirements.txt
google-cloud-bigquery>=3.14.0
google-cloud-storage>=2.14.0
db-dtypes>=1.2.0
pandas>=2.0.0
flask>=3.0.0
requests>=2.31.0
openpyxl>=3.1.0
gunicorn>=22.0.0
